Use single entrypoint/main for both Discord/Telegram
Updates are sent depending on available environment variables.
