+++ Dockerfile
@@ -6,4 +6,4 @@ RUN pip install bs4 loguru requests
 
 COPY src ./src
 
-CMD ["python3", "src/main_telegram.py"]
+CMD ["python3", "src/main.py"]
