+++ docker-compose.yaml
@@ -1,33 +1,17 @@
 version: "3.9"
 
-x-common-params: &common-params
-  image: fajfer/billy-h-scraper:latest
-  build: .
-  restart: always
-
 services:
-  billy-h-scraper-telegram:
-    <<: *common-params
-    container_name: billy-h-scraper-telegram
-    profiles:
-      - telegram
+  billy-h-scraper:
+    image: fajfer/billy-h-scraper:latest
+    build: .
+    restart: always
+    container_name: billy-h-scraper
     environment: # example
+      - MSG_DELAY=60
       - WEB_URL=https://wp.pl
       - GROUPS=@wp_pl,10101010,@global_news
       - BOT_TOKEN=000000:AAAAAABBBBBBCCCCCCC
-      - MSG_DELAY=60
-    volumes:
-      - ./history-telegram:/app/history
-  billy-h-scraper-discord:
-    <<: *common-params
-    container_name: billy-h-scraper-discord
-    profiles:
-      - discord
-    environment: # example
-      - WEB_URL=https://wp.pl
       - WEBHOOK_URL=https://discord.com/api/webhooks/.../...
       # - AVATAR_URL=  # Optional webhook avatar URL
-      - MSG_DELAY=60
     volumes:
-      - ./history-discord:/app/history
-    command: ["python3", "src/main_discord.py"]
+      - ./history:/app/history
