+++ history
@@ -1,0 +1,1 @@
+# Dummy "history" file to prevent Compose from mounting it as a directory.
