--- history-telegram
@@ -1,1 +1,0 @@
-Dummy "history" file to prevent Compose from mounting it as a directory.
